% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\docType{class}
\name{Elf-class}
\alias{Elf-class}
\alias{Elf}
\title{An S4 class to represent Elf.}
\description{
An S4 class to represent Elf.
}
\section{Slots}{

\describe{
\item{\code{classPaths}}{A list of character vectors of Java classpaths.}

\item{\code{destinationDirectory}}{A character vector of an absolute path to the
directory where the generated and compiled files should be written to.}

\item{\code{grammarFile}}{A character vector of an absolute path to a .g4 grammar
file.}

\item{\code{wunorse}}{A Java object reference to an instance of
org.rudolph.elf.Wunorse.}
}}

