% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\name{generate}
\alias{generate}
\title{generate}
\usage{
generate(self)
}
\arguments{
\item{self}{An Elf object}
}
\value{
Not meaningful.
}
\description{
Generates parser/lexer Java files using ANTLR.
}
\examples{
\dontrun{
generate(elf)
}

}
