% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\name{generateAndCompile}
\alias{generateAndCompile}
\title{generateAndCompile}
\usage{
generateAndCompile(self)
}
\arguments{
\item{self}{An Elf object}
}
\value{
Not meaningful.
}
\description{
Generates and compiles parser/lexer Java files using ANTLR and \code{javac}.
}
\examples{
\dontrun{
generateAndCompile(elf)
}

}
