% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.r
\name{getAST}
\alias{getAST}
\title{getAST}
\usage{
getAST(self, inputText)
}
\arguments{
\item{inputText}{Character vector containing text to be parsed into an AST.}
}
\value{
A nested list representing \code{inputText} parsed into an AST.
}
\description{
Generates an abstract syntax tree (AST) from a grammar. The AST returned is a
nested list.
}
\examples{
\dontrun{
ast <- getAST(rudolph, "text to be parsed")
}
}
