% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.r
\name{getGrammarMap}
\alias{getGrammarMap}
\title{getGrammarMap}
\usage{
getGrammarMap(self)
}
\value{
A named list. Rule names are the keys and the values are the
definitions.
\dontrun{
getGrammarMap(rudolph)
}
}
\description{
Returns a named list parsed from the grammar file supplied at initialization.
}
