% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\name{grammarLookup}
\alias{grammarLookup}
\title{grammarLookup}
\usage{
grammarLookup(self, ruleName)
}
\arguments{
\item{self}{A Rudolph object}

\item{ruleName}{Character vector containing name of grammar rule.}
}
\value{
A character vector representating a grammar definition (lexer or
parser).
}
\description{
Performs a lookup in the grammar file supplied at initialization. For a given
rule, returns the definition.
}
\examples{
\dontrun{
grammarLookup(rudolph, "emoticon")
}

}
