% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\docType{methods}
\name{initialize,Elf-method}
\alias{initialize,Elf-method}
\title{initialize}
\usage{
\S4method{initialize}{Elf}(.Object, destinationDirectory = character(0),
  grammarFile = character(0))
}
\arguments{
\item{destinationDirectory}{A character vector of an absolute path to the
directory where the generated and compiled files should be written to.}

\item{grammarFile}{A character vector of an absolute path to a .g4 grammar
file.}
}
\description{
Sets the working directory, initializes the JVM and Rudolph Java instance.
}
\examples{
\dontrun{
elf <- Elf(
	destinationDirectory = "/absolute/path/to/destination",
	grammarFile          = "/absolute/path/to/grammar.g4"
)
}

}
\keyword{init,}
\keyword{initialize}
