% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RudolphElf.r
\docType{methods}
\name{initialize,RudolphElf-method}
\alias{initialize,RudolphElf-method}
\title{Initialization Function}
\usage{
\S4method{initialize}{RudolphElf}(.Object)
}
\arguments{
\item{grammarFile}{.g4 grammar file}
}
\description{
Set the antlr grammar file
}
\examples{
chat <- new('RudolphElf', grammarFile='TestGrammar.g4')
show(chat)
}
\keyword{cats}

