% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{parseGrammarNameFromFile}
\alias{parseGrammarNameFromFile}
\title{parseGrammarNameFromFile}
\usage{
parseGrammarNameFromFile(grammarFile)
}
\description{
Given a file path, parse out the name of the grammar.
}
