% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\name{prettyPrint}
\alias{prettyPrint}
\title{prettyPrint}
\usage{
prettyPrint(self, ast, file)
}
\arguments{
\item{self}{A Rudolph object}

\item{ast}{A nested list representing an AST.}

\item{file}{Character vector containing file name}
}
\value{
A character vector representating composed data from the AST, or NULL
  if \code{file} is specified.
}
\description{
Takes an abstract syntax tree (AST) and naively composes a character vector
from the contents. Optionally outputs generated character vector to a named
file.
}
\examples{
\dontrun{
prettyPrint(rudolph, ast, file = "output.txt")
}

}
