% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\name{printGrammarMap}
\alias{printGrammarMap}
\title{printGrammarMap}
\usage{
printGrammarMap(self)
}
\description{
Prints a named list parsed from the grammar file supplied at initialization.
}
\details{
\dontrun{
printGrammarMap(rudolph)
}
}
