% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.r
\docType{class}
\name{Rudolph-class}
\alias{Rudolph-class}
\alias{Rudolph}
\title{An S4 class to represent an instance of Rudolph.}
\description{
An S4 class to represent an instance of Rudolph.
}
\section{Slots}{

\describe{
\item{\code{grammarFile}}{A character vector of an absolute path to a .g4 grammar
file.}

\item{\code{rootNode}}{A character vector of the root node of the grammar.}

\item{\code{rudolph}}{A Java object reference to an instance of
org.rudolph.rudolph.Rudolph.}

\item{\code{sourceDirectory}}{A character vector of an absolute path to the directory
the containing the compiled grammar files.}
}}

