% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validateFile}
\alias{validateFile}
\title{validateFile}
\usage{
validateFile(grammarFiles)
}
\arguments{
\item{grammarFiles}{A vector of absolute paths for .g4 grammar files.}
}
\description{
Checks to see if the file extension for given file is .g4 and that the file
exists.
}
