% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\docType{class}
\name{Elf-class}
\alias{Elf-class}
\alias{Elf}
\title{An S4 class to represent Elf.}
\description{
An S4 class to represent Elf.
}
\section{Slots}{

\describe{
\item{\code{classPaths}}{A list of character vectors of Java classpaths.}

\item{\code{destinationDirectory}}{A character vector of an absolute path to the
directory where the generated and compiled files should be written to.}

\item{\code{grammarFiles}}{A vector of absolute paths for .g4 grammar files.}

\item{\code{lexerPrefix}}{A character vector of the lexer file name prefix. E.g. If
the file name is "Grammar.g4" or "GrammarLexer.g4", this would be
"Grammar". If a parserPrefix is not supplied, it is assumed parserPrefix is
the same as the lexerPrefix.}

\item{\code{parserPrefix}}{A character vector of the parser file name prefix. E.g. If
the file name is "Grammar.g4" or "GrammarParser.g4", this would be
"Grammar". Not required.}

\item{\code{wunorse}}{A Java object reference to an instance of
org.rudolph.elf.Wunorse.}
}}

