% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\name{getAST}
\alias{getAST}
\title{getAST}
\usage{
getAST(self, text, file)
}
\arguments{
\item{self}{A Rudolph object}

\item{text}{Character vector containing text to be parsed into an AST.}

\item{file}{Character vector containing file name, contents of which to be
parsed into an AST.}
}
\value{
A nested list representing \code{text} or contents of \code{file}
parsed into an AST.
}
\description{
Generates an abstract syntax tree (AST) from a grammar. The AST returned is a
nested list. Only one of \code{text} or \code{file} should be specified.
}
\examples{
\dontrun{
ast <- getAST(rudolph, "text to be parsed")
}

}
