% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elf.R
\docType{methods}
\name{initialize,Elf-method}
\alias{initialize,Elf-method}
\title{initialize}
\usage{
\S4method{initialize}{Elf}(.Object, destinationDirectory = character(0),
  grammarFiles = c(), lexerName = character(0),
  parserName = character(0))
}
\arguments{
\item{destinationDirectory}{A character vector of an absolute path to the
directory where the generated and compiled files should be written to.}

\item{grammarFiles}{A vector of absolute paths for .g4 grammar files.}
}
\description{
Sets the working directory, initializes the JVM and Rudolph Java instance.
}
\examples{
\dontrun{
elf <- Elf(
	destinationDirectory = "/absolute/path/to/destination",
	grammarFiles         = c("/absolute/path/to/grammar.g4"),
	lexerName            = "grammar"
)
}

}
\keyword{init,}
\keyword{initialize}
