% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\docType{methods}
\name{initialize,Rudolph-method}
\alias{initialize,Rudolph-method}
\title{initialize}
\usage{
\S4method{initialize}{Rudolph}(.Object, grammarFiles = c(),
  rootNode = character(0), sourceDirectory = character(0),
  lexerPrefix = character(0), parserPrefix = character(0))
}
\arguments{
\item{grammarFiles}{list of file paths to .g4 grammar file. Must be an
absolute path.}

\item{rootNode}{Root node of the grammar, informs ANTLR where the grammar
rules begin.}

\item{sourceDirectory}{Path to source files. Source files are compiled Java
classes generated by ANTLR (or Elf). Must be an absolute path.}

\item{lexerPrefix}{A character vector of the lexer file name prefix. E.g. If
the file name is "Grammar.g4" or "GrammarLexer.g4", this would be
"Grammar". If a parserPrefix is not supplied, it is assumed parserPrefix is
the same as the lexerPrefix.}

\item{parserPrefix}{A character vector of the parser file name prefix. E.g.
If the file name is "Grammar.g4" or "GrammarParser.g4", this would be
"Grammar". Not required.}
}
\description{
Sets the working directory, initializes the JVM and Rudolph Java instance.
}
\examples{
\dontrun{
rudolph <- Rudolph(
	grammarFiles    = c("/absolute/path/to/grammar.g4"),
	rootNode        = "grammarroot",
	sourceDirectory = "/absolute/path/to/source",
	lexerPrefix     = "grammar"
)
}

}
\keyword{init,}
\keyword{initialize}
