% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.r
\docType{methods}
\name{initialize,Rudolph-method}
\alias{initialize,Rudolph-method}
\title{initialize}
\usage{
\S4method{initialize}{Rudolph}(.Object, grammarFile = character(0),
  rootNode = character(0), sourceDirectory = character(0))
}
\arguments{
\item{grammarFile}{File path to .g4 grammar file. Must be an absolute path.}

\item{rootNode}{Root node of the grammar, informs ANTLR where the grammar
rules begin.}

\item{sourceDirectory}{Path to source files. Source files are compiled Java
classes generated by ANTLR (or Elf). Must be an absolute path.}
}
\description{
Sets the working directory, initializes the JVM and Rudolph Java instance.
}
\examples{
\dontrun{
rudolph <- Rudolph(
	grammarFile     = "/absolute/path/to/grammar.g4",
	rootNode        = "grammarroot",
	sourceDirectory = "/absolute/path/to/source"
)
}

}
\keyword{init,}
\keyword{initialize}
