% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudolph.R
\docType{class}
\name{Rudolph-class}
\alias{Rudolph-class}
\alias{Rudolph}
\title{An S4 class to represent an instance of Rudolph.}
\description{
An S4 class to represent an instance of Rudolph.
}
\section{Slots}{

\describe{
\item{\code{grammarFiles}}{A vector of absolute paths for .g4 grammar files.}

\item{\code{rootNode}}{A character vector of the root node of the grammar.}

\item{\code{rudolph}}{A Java object reference to an instance of
org.rudolph.rudolph.Rudolph.}

\item{\code{sourceDirectory}}{A character vector of an absolute path to the directory
the containing the compiled grammar files.}

\item{\code{lexerName}}{A character vector of the lexer file name. If a parserName is
not seperately supplied, it is assumed that the parserName is the same as the
lexer name.}

\item{\code{parserName}}{A character vector of the parser file name. not required.}
}}

